backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# LangChain & LangGraph
langgraph==0.2.53
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7

# Vector store (lightweight)
faiss-cpu>=1.9.0

# Embeddings (optimized for deployment size)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers==3.2.1

# Utilities
requests==2.32.3
numpy==1.26.4
